refactor move shimmerable utility to home components and update imports
